Add tests for CSV file mode
